Tests/LibWeb/Text/input/wpt-import/encoding-detection/ja-half-width.tentative.html

<!doctype html>
<title>ja Shift_JIS half-width katakana</title>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<p>ﾊｰﾄﾞｳｪｱﾊｰﾄﾞｳｪｱﾊｰﾄﾞｳｪｱﾊｰﾄﾞｳｪｱﾊｰﾄﾞｳｪｱ</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "Shift_JIS", 'Expected Shift_JIS');
    }, "Check detection result");
      done();
};
</script>
