third_party/blink/web_tests/external/wpt/encoding-detection/ja-half-width.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Check detection result
  assert_equals: Expected Shift_JIS expected "Shift_JIS" but got "gb18030"
Harness: the test ran to completion.
